refactor: Q클래스 import 출처 재지정
