+++ src/main/java/umc/stockoneqback/board/infra/query/BoardListQueryRepositoryImpl.java
@@ -11,7 +11,7 @@ import org.springframework.transaction.annotation.Transactional;
 import umc.stockoneqback.board.domain.BoardSearchType;
 import umc.stockoneqback.board.dto.response.BoardList;
 import umc.stockoneqback.board.dto.response.CustomBoardListResponse;
-import umc.stockoneqback.board.infra.query.dto.response.QBoardList;
+import umc.stockoneqback.board.dto.response.QBoardList;
 import umc.stockoneqback.global.base.Status;
 
 import java.util.List;
