+++ src/main/java/umc/stockoneqback/share/infra/query/ShareListQueryRepositoryImpl.java
@@ -12,13 +12,13 @@ import umc.stockoneqback.global.base.Status;
 import umc.stockoneqback.share.domain.model.Category;
 import umc.stockoneqback.share.domain.model.ShareSearchType;
 import umc.stockoneqback.share.dto.response.CustomShareListPage;
+import umc.stockoneqback.share.dto.response.QShareList;
 import umc.stockoneqback.share.dto.response.ShareList;
-import umc.stockoneqback.share.infra.query.dto.QShareList;
 
 import java.util.List;
 
 import static umc.stockoneqback.business.domain.QBusiness.business;
-import static umc.stockoneqback.share.domain.QShare.share;
+import static umc.stockoneqback.share.domain.model.QShare.share;
 
 @Transactional(readOnly = true)
 @RequiredArgsConstructor
