+++ src/main/java/umc/stockoneqback/user/infra/query/UserFindQueryRepositoryImpl.java
@@ -7,7 +7,7 @@ import org.springframework.transaction.annotation.Transactional;
 import umc.stockoneqback.global.base.Status;
 import umc.stockoneqback.user.domain.search.UserSearchType;
 import umc.stockoneqback.user.dto.FindManager;
-import umc.stockoneqback.user.infra.query.dto.QFindManager;
+import umc.stockoneqback.user.dto.QFindManager;
 
 import java.util.List;
