+++ src/test/java/umc/stockoneqback/business/controller/BusinessProductApiControllerTest.java
@@ -11,9 +11,9 @@ import umc.stockoneqback.common.ControllerTest;
 import umc.stockoneqback.fixture.ProductFixture;
 import umc.stockoneqback.global.exception.BaseException;
 import umc.stockoneqback.global.exception.GlobalErrorCode;
-import umc.stockoneqback.product.exception.ProductErrorCode;
 import umc.stockoneqback.product.dto.response.GetTotalProductResponse;
 import umc.stockoneqback.product.dto.response.SearchProductOthersResponse;
+import umc.stockoneqback.product.exception.ProductErrorCode;
 import umc.stockoneqback.user.exception.UserErrorCode;
 
 import java.util.ArrayList;
