+++ src/test/java/umc/stockoneqback/business/infra/query/BusinessListQueryRepositoryImplTest.java
@@ -9,8 +9,8 @@ import umc.stockoneqback.business.domain.BusinessRepository;
 import umc.stockoneqback.business.dto.response.BusinessList;
 import umc.stockoneqback.common.RepositoryTest;
 import umc.stockoneqback.field.domain.model.Company;
-import umc.stockoneqback.field.domain.repository.CompanyRepository;
 import umc.stockoneqback.field.domain.model.Store;
+import umc.stockoneqback.field.domain.repository.CompanyRepository;
 import umc.stockoneqback.field.domain.repository.StoreRepository;
 import umc.stockoneqback.global.base.RelationStatus;
 import umc.stockoneqback.user.domain.User;
